+++ src/services/Search.php
@@ -13,6 +13,7 @@ use craft\base\FieldInterface;
 use craft\base\MemoizableArray;
 use craft\db\Query;
 use craft\db\Table;
+use craft\elements\db\ElementQuery;
 use craft\errors\SiteNotFoundException;
 use craft\events\SearchEvent;
 use craft\helpers\ArrayHelper;
@@ -24,6 +25,7 @@ use craft\search\SearchQuery;
 use craft\search\SearchQueryTerm;
 use craft\search\SearchQueryTermGroup;
 use yii\base\Component;
+use yii\db\Expression;
 use yii\db\Schema;
 
 /**
@@ -204,11 +206,12 @@ class Search extends Component
      *
      * @param int[] $elementIds The list of element IDs to filter by the search query.
      * @param string|array|SearchQuery $searchQuery The search query (either a string or a SearchQuery instance)
-     * @param bool $scoreResults Whether to order the results based on how closely they match the query.
+     * @param bool $scoreResults Whether to order the results based on how closely they match the query. (No longer checked.)
      * @param int|int[]|null $siteId The site ID(s) to filter by.
      * @param bool $returnScores Whether the search scores should be included in the results. If true, results will be returned as `element ID => score`.
      * @param FieldInterface[]|null $customFields The custom fields involved in the query.
      * @return array The filtered list of element IDs.
+     * @deprecated in 3.7.14. Use [[searchElements()]] instead.
      */
     public function filterElementIdsByQuery(
         array $elementIds,
@@ -218,6 +221,46 @@ class Search extends Component
         bool $returnScores = false,
         ?array $customFields = null
     ): array {
+        $scoredResults = $this->_searchElements(null, $elementIds, $searchQuery, $siteId, $customFields);
+        return $returnScores ? $scoredResults : array_keys($scoredResults);
+    }
+
+    /**
+     * Searches for elements that match the given element query.
+     *
+     * @param ElementQuery $elementQuery The element query being executed
+     * @return array The filtered list of element IDs.
+     * @since 3.7.14
+     */
+    public function searchElements(ElementQuery $elementQuery): array
+    {
+        return $this->_searchElements($elementQuery, null, $elementQuery->search, $elementQuery->siteId, $elementQuery->customFields);
+    }
+
+    /**
+     * Filters a list of element IDs by a given search query.
+     *
+     * @param ElementQuery|null $elementQuery
+     * @param int[]|null $elementIds
+     * @param string|array|SearchQuery $searchQuery
+     * @param int|int[]|null $siteId
+     * @param FieldInterface[]|null $customFields
+     * @return array
+     */
+    private function _searchElements(
+        ?ElementQuery $elementQuery,
+        ?array $elementIds,
+        $searchQuery,
+        $siteId,
+        ?array $customFields
+    ): array {
+        if ($elementQuery !== null) {
+            $elementQuery = (clone $elementQuery)
+                ->search(null)
+                ->offset(null)
+                ->limit(null);
+        }
+
         if (is_string($searchQuery)) {
             $searchQuery = new SearchQuery($searchQuery, Craft::$app->getConfig()->getGeneral()->defaultSearchTermOptions);
         } else if (is_array($searchQuery)) {
@@ -231,7 +274,7 @@ class Search extends Component
         // Fire a 'beforeSearch' event
         if ($this->hasEventHandlers(self::EVENT_BEFORE_SEARCH)) {
             $this->trigger(self::EVENT_BEFORE_SEARCH, new SearchEvent([
-                'elementIds' => $elementIds,
+                'elementQuery' => $elementQuery,
                 'query' => $searchQuery,
                 'siteId' => $siteId,
             ]));
@@ -258,94 +301,58 @@ class Search extends Component
         // Get where clause from tokens, bail out if no valid query is there
         $where = $this->_getWhereClause($siteId, $customFields);
 
-        if ($where === false || empty($where)) {
+        if (empty($where)) {
             return [];
         }
 
-        $db = Craft::$app->getDb();
+        $query = (new Query())
+            ->from([Table::SEARCHINDEX])
+            ->where(new Expression($where));
 
         if ($siteId !== null) {
-            if (is_array($siteId)) {
-                $where .= sprintf(' AND %s IN (%s)',
-                    $db->quoteColumnName('siteId'),
-                    implode(',', $siteId)
-                );
-            } else {
-                $where .= sprintf(' AND %s = %s', $db->quoteColumnName('siteId'), $db->quoteValue($siteId));
-            }
+            $query->andWhere(['siteId' => $siteId]);
         }
 
-        // Begin creating SQL
-        $sql = sprintf('SELECT * FROM %s WHERE %s', $db->quoteTableName(Table::SEARCHINDEX), $where);
-
-        // Append elementIds to QSL
-        if (!empty($elementIds)) {
-            $sql .= sprintf(' AND %s IN (%s)',
-                $db->quoteColumnName('elementId'),
-                implode(',', $elementIds)
-            );
+        if ($elementQuery !== null) {
+            $query->andWhere([
+                'elementId' => $elementQuery->select(['elements.id']),
+            ]);
+        } else if (!empty($elementIds)) {
+            $query->andWhere(['elementId' => $elementIds]);
         }
 
         // Execute the sql
-        $results = $db->createCommand($sql)->queryAll();
+        $results = $query->all();
 
-        // Are we scoring the results?
-        if ($scoreResults) {
-            $scoresByElementId = [];
+        // Score the results
+        $scoresByElementId = [];
 
-            // Loop through results and calculate score per element
-            foreach ($results as $row) {
-                $elementId = $row['elementId'];
-                $score = $this->_scoreRow($row, $siteId);
-
-                if (!isset($scoresByElementId[$elementId])) {
-                    $scoresByElementId[$elementId] = $score;
-                } else {
-                    $scoresByElementId[$elementId] += $score;
-                }
-            }
-
-            // Sort found elementIds by score
-            arsort($scoresByElementId);
-
-            // Fire an 'afterSearch' event
-            if ($this->hasEventHandlers(self::EVENT_AFTER_SEARCH)) {
-                $this->trigger(self::EVENT_AFTER_SEARCH, new SearchEvent([
-                    'elementIds' => array_keys($scoresByElementId),
-                    'query' => $searchQuery,
-                    'siteId' => $siteId,
-                    'results' => $results,
-                ]));
-            }
-
-            if ($returnScores) {
-                return $scoresByElementId;
-            }
-
-            // Just return the ordered element IDs
-            return array_keys($scoresByElementId);
-        }
-
-        // Don't apply score, just return the IDs
-        $elementIds = [];
-
+        // Loop through results and calculate score per element
         foreach ($results as $row) {
-            $elementIds[] = $row['elementId'];
+            $elementId = $row['elementId'];
+            $score = $this->_scoreRow($row, $siteId);
+
+            if (!isset($scoresByElementId[$elementId])) {
+                $scoresByElementId[$elementId] = $score;
+            } else {
+                $scoresByElementId[$elementId] += $score;
+            }
         }
 
-        $elementIds = array_unique($elementIds);
+        arsort($scoresByElementId);
 
         // Fire an 'afterSearch' event
         if ($this->hasEventHandlers(self::EVENT_AFTER_SEARCH)) {
             $this->trigger(self::EVENT_AFTER_SEARCH, new SearchEvent([
-                'elementIds' => $elementIds,
+                'elementQuery' => $elementQuery,
+                'elementIds' => array_keys($scoresByElementId),
                 'query' => $searchQuery,
                 'siteId' => $siteId,
                 'results' => $results,
             ]));
         }
 
-        return $elementIds;
+        return $scoresByElementId;
     }
 
     /**